requirements/common.pip.txt
Django==1.5.5
Markdown==2.3.1
djangorestframework==2.3.8
